+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 0.4.6
+Version 0.5.0
 
 Changes:
 
@@ -6,7 +6,7 @@ Changes:
 - Cleaned up some code
 - Added some error handling
 - Added some better filesystem handling
-    - Okay, it doesn't really work that well
+    - The makefile now sets the project directory via a macro
 - Changed .gitignore
 - Updated make clean to also clean out binaries
 - Changed the `Camera` class to allow for three different types of parallel types: 0 (cpu), 1 (cpu, multi), 2 (gpu)
@@ -19,3 +19,5 @@ Changes:
 - `cuda_metric.cuh` has been created and contains a CUDA-suitable metric, which takes calls from the config file.
 - Updated `makefile` so it is more concise and so that the CUDA code only compiles if nvcc is installed.
     - It also creates build, bin folders if necessary.
+- Implemented PNG, JPG handling with the `stb_image_write.h` and `stb_image.h`
+- Updated README.md.